refactor: rename internal api client to managementapi
it seems to me clearer like that this package allows to interact with the Api management
