+++ controllers/apidefinition_controller_start_stop_test.go
@@ -24,7 +24,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	gio "github.com/gravitee-io/gravitee-kubernetes-operator/api/v1alpha1"
-	apim "github.com/gravitee-io/gravitee-kubernetes-operator/internal/apim/client"
+	managementapi "github.com/gravitee-io/gravitee-kubernetes-operator/internal/apim/managementapi"
 	"github.com/gravitee-io/gravitee-kubernetes-operator/test"
 )
 
@@ -115,7 +115,7 @@ var _ = Describe("API Definition Controller", func() {
 			}, timeout, interval).Should(BeTrue())
 
 			By("Call rest API and expect STOPPED state")
-			apimClient := apim.NewClient(ctx, managementContextFixture, httpClient)
+			apimClient := managementapi.NewClient(ctx, managementContextFixture, httpClient)
 			Eventually(func() bool {
 				api, apiErr := apimClient.GetByCrossId(updatedApiDefinition.Status.CrossID)
