+++ internal/apim/apis/types.go
@@ -7,20 +7,24 @@ import (
 
 	"github.com/go-logr/logr"
 	gio "github.com/gravitee-io/gravitee-kubernetes-operator/api/v1alpha1"
-	apim "github.com/gravitee-io/gravitee-kubernetes-operator/internal/apim/client"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	managementapi "github.com/gravitee-io/gravitee-kubernetes-operator/internal/apim/managementapi"
+	k8s "sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+)
+
+const (
+	requestTimeoutSeconds = 5
 )
 
 type Delegate struct {
 	ctx               context.Context
 	managementContext *gio.ManagementContext
-	apimClient        *apim.Client
-	k8sClient         client.Client
+	apimClient        *managementapi.Client
+	k8sClient         k8s.Client
 	log               logr.Logger
 }
 
-func NewDelegate(ctx context.Context, client client.Client) *Delegate {
+func NewDelegate(ctx context.Context, client k8s.Client) *Delegate {
 	log := log.FromContext(ctx)
 
 	return &Delegate{
@@ -36,7 +40,7 @@ func (d *Delegate) SetManagementContext(managementContext *gio.ManagementContext
 	d.managementContext = managementContext
 
 	httpClient := http.Client{Timeout: requestTimeoutSeconds * time.Second}
-	d.apimClient = apim.NewClient(d.ctx, d.managementContext, httpClient)
+	d.apimClient = managementapi.NewClient(d.ctx, d.managementContext, httpClient)
 }
 
 func (d *Delegate) IsConnectedToManagementApi() bool {
