+++ internal/apim/managementapi/managementapi.go
@@ -1,4 +1,4 @@
-package apim
+package managementapi
 
 import (
 	"context"
